systemtap: Use correct formatters/types
Try to avoid casting variables to potentially smaller types to fit
defined probes. This can truncate values and lead to wrong results.

Also make sure that signedness matches.

Since GType can be even 128 bit on CHERI architecture, cast all these
various types used based on platform to uintmax_t which SystemTap
properly processes.

+++ glib/gmem.c
@@ -98,7 +98,7 @@ g_malloc (gsize n_bytes)
       gpointer mem;
 
       mem = malloc (n_bytes);
-      TRACE (GLIB_MEM_ALLOC((void*) mem, (unsigned int) n_bytes, 0, 0));
+      TRACE (GLIB_MEM_ALLOC ((void *) mem, n_bytes, 0, 0));
       if (mem)
 	return mem;
 
@@ -106,7 +106,7 @@ g_malloc (gsize n_bytes)
                G_STRLOC, n_bytes);
     }
 
-  TRACE(GLIB_MEM_ALLOC((void*) NULL, (int) n_bytes, 0, 0));
+  TRACE (GLIB_MEM_ALLOC ((void *) NULL, n_bytes, 0, 0));
 
   return NULL;
 }
@@ -131,7 +131,7 @@ g_malloc0 (gsize n_bytes)
       gpointer mem;
 
       mem = calloc (1, n_bytes);
-      TRACE (GLIB_MEM_ALLOC((void*) mem, (unsigned int) n_bytes, 1, 0));
+      TRACE (GLIB_MEM_ALLOC ((void *) mem, n_bytes, 1, 0));
       if (mem)
 	return mem;
 
@@ -139,7 +139,7 @@ g_malloc0 (gsize n_bytes)
                G_STRLOC, n_bytes);
     }
 
-  TRACE(GLIB_MEM_ALLOC((void*) NULL, (int) n_bytes, 1, 0));
+  TRACE (GLIB_MEM_ALLOC ((void *) NULL, n_bytes, 1, 0));
 
   return NULL;
 }
@@ -169,7 +169,7 @@ g_realloc (gpointer mem,
   if (G_LIKELY (n_bytes))
     {
       newmem = realloc (mem, n_bytes);
-      TRACE (GLIB_MEM_REALLOC((void*) newmem, (void*)mem, (unsigned int) n_bytes, 0));
+      TRACE (GLIB_MEM_REALLOC ((void *) newmem, (void *) mem, n_bytes, 0));
       if (newmem)
 	return newmem;
 
@@ -320,7 +320,7 @@ g_try_malloc (gsize n_bytes)
   else
     mem = NULL;
 
-  TRACE (GLIB_MEM_ALLOC((void*) mem, (unsigned int) n_bytes, 0, 1));
+  TRACE (GLIB_MEM_ALLOC ((void *) mem, n_bytes, 0, 1));
 
   return mem;
 }
@@ -375,7 +375,7 @@ g_try_realloc (gpointer mem,
       free (mem);
     }
 
-  TRACE (GLIB_MEM_REALLOC((void*) newmem, (void*)mem, (unsigned int) n_bytes, 1));
+  TRACE (GLIB_MEM_REALLOC ((void *) newmem, (void *) mem, n_bytes, 1));
 
   return newmem;
 }
@@ -647,7 +647,7 @@ g_aligned_alloc (gsize n_blocks,
 
   if (G_UNLIKELY (real_size == 0))
     {
-      TRACE(GLIB_MEM_ALLOC((void*) NULL, (int) real_size, 0, 0));
+      TRACE (GLIB_MEM_ALLOC ((void *) NULL, real_size, 0, 0));
       return NULL;
     }
 
@@ -685,7 +685,7 @@ g_aligned_alloc (gsize n_blocks,
 # error "This platform does not have an aligned memory allocator."
 #endif
 
-  TRACE (GLIB_MEM_ALLOC((void*) res, (unsigned int) real_size, 0, 0));
+  TRACE (GLIB_MEM_ALLOC ((void *) res, real_size, 0, 0));
   if (res)
     return res;
